+++ CodeConverter/Util/ExpressionSyntaxExtensions.cs
@@ -26,11 +26,11 @@ namespace ICSharpCode.CodeConverter.Util
             return VBasic.SyntaxFactory.ArgumentList(VBasic.SyntaxFactory.SeparatedList(argExpressions.Select(e => (VBSyntax.ArgumentSyntax) VBasic.SyntaxFactory.SimpleArgument(e))));
         }
 
-        public static ArgumentListSyntax CreateCsArgList<T>(this IEnumerable<T> argExpressions, params SyntaxKind[] refTokenKinds) where T : ExpressionSyntax
+        public static ArgumentListSyntax CreateCsArgList<T>(this IEnumerable<T> argExpressions, params SyntaxKind?[] refTokenKinds) where T : ExpressionSyntax
         {
             return SyntaxFactory.ArgumentList(SyntaxFactory.SeparatedList(argExpressions.Select((e, i) => {
                 var arg = SyntaxFactory.Argument(e);
-                if (i < refTokenKinds.Length) arg = arg.WithRefKindKeyword(SyntaxFactory.Token(refTokenKinds[i]));
+                if (i < refTokenKinds.Length && refTokenKinds[i].HasValue) arg = arg.WithRefKindKeyword(SyntaxFactory.Token(refTokenKinds[i].Value));
                 return arg;
             })));
         }